Add hook to work around AUR issue
In VS Code, workspace-level settings need to live in a subdirectory
but the AUR doesn’t accept submissions with subdirectories.

In order to still allow VS Code users to collaborate, add a hook that
creates a .vscode.untar-me.tar.xz tarball, which can be checked into
a VCS and unpacked later.